Update amedata.py: check for database print
I was thinking about logging the prints in the check for database function since they are always printed each time you call the class. However, I think that is unnecessary since if it is not available, it will try to download it until it finds it. Therefore I just deleted it.